CustomerServiceChatbot/requirements.txt
# langchain==0.0.339
# python-dotenv==1.0.0
# streamlit==1.22.0
# tiktoken==0.4.0
# faiss-cpu==1.7.4
# protobuf~=3.19.0
# sentence-transformer== 2.2.2
langchain
python-dotenv
streamlit
tiktoken
faiss-cpu
protobuf
sentence-transformer